+++ crates/lune-std/src/globals/mod.rs
@@ -1,5 +1,6 @@
 pub mod g_table;
 pub mod print;
 pub mod require;
+pub mod script;
 pub mod version;
 pub mod warn;